Revert "Moves root volume resize logic to cfn-init"
This reverts commit 306a10aacc1ced6e017dca68d2e5986321f04752.

We can keep this logic in userdata for now. Currently, this patch adds
the ability to extend the root volume only at launch, so it makes more
sense to keep the growpart and pvresize logic together.

Support for extending the root volume during a stack update can be
done separately.