Removes install of cfn-utils from userdata
Originally, this userdata was designed to work on generic AMIs from RH or CentOS.
Those AMIs do not include the cfn-utils, so we installed them. However, those utils
only support python2 at the moment. We have since moved to python3 primarily, and
we only test using SPEL (which has the cfn-utils pre-installed), so we did not
notice that the cfn-util install was no longer effective.

Peeling back the onion, python2 in EL6/7 does not have the ensurepip module. So
to install pip in python2 so that we can install the cfn-utils, we would have to
re-add the GetPipUrl parameter to install pip the old way.

And while AWS does provide an RPM installer for cfn-utils, the RPM depends on EPEL
packages, which would require new parameters again for the plain RH/CentOS AMIs
(to install the epel release pkg and enable the repo).

All sounds a bit lame to me. Figuring that for now, until cfn-utils supports python3,
easier just to remove the cfn-util install since it is not working anyway. To use
these templates then, we will have a requirement that the cfn-utils be pre-installed
in the AMI.